app/requirements.txt
asgiref==3.6.0
attrs==22.2.0
autopep8==2.0.1
cx-Oracle==8.3.0
Django==4.1.7
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
drf-spectacular==0.25.1
inflection==0.5.1
jsonschema==4.17.3
Markdown==3.4.1
mysqlclient==2.1.1
pycodestyle==2.10.0
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
uritemplate==4.1.1
